added swagger documentation
