--- back-end/app/Http/Controllers/AuthController.php
@@ -1,243 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\User;
-use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
-
-class AuthController extends Controller
-{
-
-    /**
-     * @OA\Post(
-     *     path="api/auth/register",
-     *     tags={"Authentication"},
-     *     summary="Register",
-     *     description="Register a new user",
-     *     @OA\RequestBody(
-     *         required=true,
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="name",
-     *                 type="string",
-     *                 example="John Doe"
-     *             ),
-     *             @OA\Property(
-     *                 property="email",
-     *                 type="string",
-     *                 format="email",
-     *                 example="user@example.com"
-     *             ),
-     *             @OA\Property(
-     *                 property="password",
-     *                 type="string",
-     *                 format="password",
-     *                 example="password123"
-     *             )
-     *         )
-     *     ),
-     *     @OA\Response(
-     *         response=201,
-     *         description="User created successfully",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="User created successfully"
-     *             ),
-     *             @OA\Property(
-     *                 property="token",
-     *                 type="string",
-     *                 example="random string token"
-     *             )
-     *         )
-     *     ),
-     *     @OA\Response(
-     *         response=400,
-     *         description="Bad Request",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="The given data was invalid."
-     *             )
-     *         )
-     *     )
-     * )
-     */
-    public function register(Request $request): JsonResponse
-    {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required|email',
-            'password' => 'required|string'
-        ]);
-
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password)
-        ]);
-
-        return response()->json([
-            'message' => 'User created successfully',
-            'token' => $user->createToken('token')->plainTextToken
-        ]);
-    }
-    /**
-     * @OA\Post(
-     *     path="api/auth/login",
-     *     tags={"Authentication"},
-     *     summary="Login",
-     *     description="Log in",
-     *     @OA\RequestBody(
-     *         required=true,
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="email",
-     *                 type="string",
-     *                 format="email",
-     *                 example="user@example.com"
-     *             ),
-     *             @OA\Property(
-     *                 property="password",
-     *                 type="string",
-     *                 format="password",
-     *                 example="password123"
-     *             )
-     *         )
-     *     ),
-     *     @OA\Response(
-     *         response=200,
-     *         description="The login was successful",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="Logged in successfully"
-     *             ),
-     *             @OA\Property(
-     *                 property="token",
-     *                 type="string",
-     *                 example="random string token"
-     *             )
-     *         )
-     *     ),
-     *     @OA\Response(
-     *         response=401,
-     *         description="Login failed",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="Invalid credentials"
-     *             )
-     *         )
-     *     )
-     * )
-     */
-    public function login(Request $request): JsonResponse
-    {
-        $request->validate([
-            'email' => 'required|email',
-            'password' => 'required|string'
-        ]);
-
-        $user = User::where('UserEmail', $request->email)->first();
-        if (!$user || !Hash::check($request->password, $user->UserPassword)) {
-            return response()->json([
-                'message' => 'Invalid credentials',
-            ], 401);
-        }
-
-        return response()->json([
-            'message' => 'Logged in successfully',
-            'token' => $user->createToken('token')->plainTextToken
-        ]);
-    }
-
-    /**
-     * @OA\Post(
-     *     path="api/auth/logout",
-     *     tags={"Authentication"},
-     *     summary="Logout",
-     *     description="Log out the current user",
-     *     security={{"sanctum":{}}},
-     *     @OA\Response(
-     *         response=200,
-     *         description="Logged out successfully",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="Logged out successfully"
-     *             )
-     *         )
-     *     ),
-     *     @OA\Response(
-     *         response=401,
-     *         description="Unauthorized",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="Unauthorized"
-     *             )
-     *         )
-     *     )
-     * )
-     */
-    public function logout(Request $request): JsonResponse
-    {
-        $request->user('sanctum')->currentAccessToken()->delete();
-
-
-        return response()->json(['message' => 'Logged out successfully']);
-    }
-
-    /**
-     * @OA\Get(
-     *     path="api/auth/check",
-     *     tags={"Authentication"},
-     *     summary="Check Authorization",
-     *     description="Check if the user is authenticated",
-     *     security={{"sanctum":{}}},
-     *     @OA\Response(
-     *         response=200,
-     *         description="Authorized",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="Authorized"
-     *             )
-     *         )
-     *     ),
-     *     @OA\Response(
-     *         response=401,
-     *         description="Unauthorized",
-     *         @OA\JsonContent(
-     *             @OA\Property(
-     *                 property="message",
-     *                 type="string",
-     *                 example="Unauthorized"
-     *             )
-     *         )
-     *     )
-     * )
-     */
-    public function check(Request $request): JsonResponse
-    {
-        $user = auth('sanctum')->user();
-
-        if (!$user) {
-            return response()->json(['message' => 'Unauthorized'], 401);
-        }
-
-        return response()->json([
-            'message' => 'Authorized',
-        ]);
-    }
-}
